+++ src/routes/api/yahoo-finance/index.ts
