+++ src/routes/exchange-rates/index.svelte
@@ -1,6 +1,6 @@
 <script context="module">
 	export async function load({ fetch }) {
-		const response = await fetch('./api/exchange-rates', {
+		const body = await fetch('./api/yahoo-finance', {
 			method: 'POST',
 			body: JSON.stringify({
 				from: new Date().getTime() / 1000 - 7 * 24 * 60 * 60,
@@ -10,8 +10,7 @@
 			headers: {
 				'Content-Type': 'application/json'
 			}
-		});
-		const body = await response.json();
+		}).then((response) => response.json());
 		return {
 			props: {
 				data: body.data
@@ -21,29 +20,20 @@
 </script>
 
 <script lang="ts">
-	import Card from '$lib/components/card.svelte';
-	import LineChart from '$components/single-line-chart.svelte';
 	import { currentPageStore } from '$stores/current-page-store';
 	import { onMount, onDestroy } from 'svelte';
+	import CardList from '$components/card-list.svelte';
 
 	export let container = 'default-container';
 	export let data = [];
-	let graphData = [];
 
 	onMount(() => {
-		graphData = data;
 		currentPageStore.set('Exchange-Rates');
 	});
 
 	onDestroy(() => {});
 </script>
 
-<div class="{container} grid gap-4 grid-cols-1 xl:grid-cols-2 2xl:grid-cols-2 3xl:grid-cols-3">
-	{#each graphData as lineData}
-		<Card name={lineData.symbol}>
-			<LineChart {lineData} />
-		</Card>
-	{:else}
-		<h1>Loading...</h1>
-	{/each}
+<div class="{container}">
+	<CardList {data}></CardList>
 </div>
